// src/app/generated/caster-api/model/vlan.ts
/*
Copyright 2021 Carnegie Mellon University. All Rights Reserved. 
 Released under a MIT (SEI)-style license. See LICENSE.md in the project root for license information.
*/

/**
 * Caster API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Vlan { 
    id?: string;
    poolId?: string;
    partitionId?: string | null;
    vlanId?: number;
    inUse?: boolean;
    tag?: string | null;
    reserved?: boolean;
    reservedEditable?: boolean;
}